apps/runtime/pyproject.toml
[tool.poetry]
name = "flowos-runtime"
version = "0.1.0"
description = "FlowOS LangGraph execution runtime"
authors = ["FlowOS"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
langgraph = "^0.1.0"
langchain-core = "^0.2.0"
langchain-anthropic = "^0.1.0"
langchain-openai = "^0.1.0"
supabase = "^2.4.0"
httpx = "^0.27.0"
pydantic = "^2.7.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
